rename src/index to ed
